Clean legacy IO80211SkywalkInterface  import.
