+++ include/Airport/Apple80211.h
@@ -29,9 +29,6 @@
 #include "IO80211Interface.h"
 #include "IO80211VirtualInterface.h"
 #include "IO80211P2PInterface.h"
-#if __IO80211_TARGET >= __MAC_10_15
-#include "IO80211SkywalkInterface.h"
-#endif
 #endif /* IO80211FAMILY_V2 */
 
 #endif /* Apple80211_h */
